Updating tests to create the indexes in the db before running.
