+++ test/@data/drop-and-recreate.js
@@ -5,7 +5,7 @@ var findItemsToCreate = require('./find-items-to-create');
 var insertData = require('./insert-data');
 var mongo = require('../../src/mongo/index');
 var permissions = require('../../src/permissions');
-
+var createIndexes = require('../../utilities/create-indexes/create');
 module.exports = function dropAndRecreate(callback) {
     async.waterfall([
         function (cb) {
@@ -15,10 +15,12 @@ module.exports = function dropAndRecreate(callback) {
         },
         dropExistingData,
         initialiseRoles,
+        createIndexes,
         findItemsToCreate,
         insertData
     ], callback);
 };
+
 function initialiseRoles(callback) {
     permissions.initialise({}, function (err) {
         return callback(err);
